rename xr.plot.plot module since it shadows the xr.plot.plot method
