+++ xarray/plot/__init__.py
@@ -1,6 +1,24 @@
+"""
+Use this module directly:
+    import xarray.plot as xplt
+
+Or use the methods on a DataArray or Dataset:
+    DataArray.plot._____
+    Dataset.plot._____
+"""
+from .dataarray_plot import (
+    contour,
+    contourf,
+    hist,
+    imshow,
+    line,
+    pcolormesh,
+    plot,
+    step,
+    surface,
+)
 from .dataset_plot import scatter
 from .facetgrid import FacetGrid
-from .plot import contour, contourf, hist, imshow, line, pcolormesh, plot, step, surface
 
 __all__ = [
     "plot",
